TS対応
