+++ package.json
@@ -11,7 +11,13 @@
   "type": "module",
   "dependencies": {
     "@aws-sdk/client-s3": "^3.758.0",
+    "@types/express": "^5.0.0",
     "dotenv": "^16.4.7",
-    "express": "^4.21.2"
+    "express": "^4.21.2",
+    "typescript": "^5.8.2"
+  },
+  "devDependencies": {
+    "@types/node": "^22.13.10",
+    "tsx": "^4.19.3"
   }
 }
