--- src/services/s3-client.js
@@ -1,15 +1,0 @@
-import { S3Client } from "@aws-sdk/client-s3";
-import dotenv from "dotenv";
-
-dotenv.config();
-
-// S3 を操作するためのインスタンスを生成
-const s3Client = new S3Client({
-  credentials: {
-    accessKeyId: process.env.AWS_ACCESS_KEY,
-    secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
-  },
-  region: process.env.AWS_S3_REGION,
-});
-
-export default s3Client;
